"ads2_m8_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ assignments/m8/1. Shortest Paths/Solution.java
@@ -37,9 +37,11 @@ public final class Solution {
     EdgeWeightedDigraph eg = new EdgeWeightedDigraph(vertices);
     while (edges > 0) {
       String[] token = scan.nextLine().split(" ");
+      // String a = token[0];
+      // String b = token[1];
       int first = sol.getIndex(token[0]);
       int second = sol.getIndex(token[1]);
-      int weight = Integer.parseInt(token[2]);
+      double weight = Double.parseDouble(token[2]);
       DirectedEdge e = new DirectedEdge(first, second, weight);
       eg.addEdge(e);
       edges--;
